Added 4th table as template.

+++ src/com/dozingcatsoftware/vectorpinball/tables/Field3Delegate.java
@@ -163,7 +163,7 @@ public class Field3Delegate extends BaseFieldDelegate {
             // rollover groups increment field multiplier when all rollovers are activated, also reset to inactive
             rolloverGroup.setAllRolloversActivated(false);
             field.getGameState().incrementScoreMultiplier();
-            field.showGameMessage(field.getGameState().getScoreMultiplier() + "x Multiplier", 1500);
+            field.showGameMessage(((int)field.getGameState().getScoreMultiplier()) + "x Multiplier", 1500);
         }
     }
